requirements.txt
pandas>=0.23.4
intel-numpy
intel-scipy
tqdm
seaborn
